requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
djangorestframework==3.12.4
fcm-django==0.3.10
idna==2.10
pyfcm==1.4.9
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
